Initial tooljs commit:
The tooljs project includes utilities to make jake files look more like
make files.  One of the goals is consistant behavior between linux, mac
and windows.  Command line utilities that have differences between the
systems will be duplicated.  The project will attempt to be as
asynchronous as possible.

The current features are:

- finddir returns files that match wildard recusively

- support for cp, rm, mkdir, rmdir, echo, exec

- async with callbacks

- chaining style creation of execution lists, which is gear for the
  nodejs async module